CID-220 removed integrationapi.enabled flag since integration runs always by ihub. Removed unused integration_api functions
