+++ helm/leanix-k8s-connector/templates/cronjob.yaml
@@ -65,11 +65,8 @@ spec:
               value: "{{ .Values.args.processingMode }}"
             - name: BLACKLIST_NAMESPACES
               value: "{{ .Values.args.blacklistNamespaces | join ", " }}"
-            {{- if .Values.integrationApi.enabled }}
             - name: INTEGRATION_API_DATASOURCENAME
               value: "{{ .Values.integrationApi.datasourceName }}"
-            - name: INTEGRATION_API_ENABLED
-              value: "true"
             - name: INTEGRATION_API_FQDN
               value: "{{ .Values.integrationApi.fqdn }}"
             - name: INTEGRATION_API_TOKEN
@@ -77,7 +74,6 @@ spec:
                 secretKeyRef:
                   name: "{{ .Values.integrationApi.secretName }}"
                   key: token
-            {{- end }}
             {{- range $key, $val := .Values.args.additionalEnv }}
             - name: {{ $key }}
               value: {{ $val | quote }}
